+++ app/src/models/auth/jwtService.js
@@ -4,7 +4,7 @@ import jwt from "jsonwebtoken";
 class JwtToken {
   static async createAccessToken(id) {
     const accessToken = jwt.sign(id, process.env.ACCESS_TOKEN_SECRET, {
-      expiresIn: "5m",
+      expiresIn: "30m",
     })
     return accessToken;
   }
@@ -21,12 +21,12 @@ class JwtToken {
       return { error: 'Unauthorized', message: err.message, statusCode: 401 };
     }
   }
-  static async verifyAccessToken(token) {
-    try {
-      return jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
-    } catch (err) {
-      return { error: 'Unauthorized', message: err.message, statusCode: 401 };
-    }
-  }
+  // static async verifyAccessToken(token) {
+  //   try {
+  //     return jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+  //   } catch (err) {
+  //     return { error: 'Unauthorized', message: err.message, statusCode: 401 };
+  //   }
+  // }
 }
 export default JwtToken;
