+++ app/src/routes/auth/authCtrl.js
@@ -1,13 +1,13 @@
 "use strict";
 
 import Auth from "../../models/auth/authService.js"
-const output = {
-  check: async (req, res) => {
-    const auth = new Auth(req);
-    const response = await auth.check();
-    return res.json(response).status(response.statusCode);
-  }
-}
+// const output = {
+//   check: async (req, res) => {
+//     const auth = new Auth(req);
+//     const response = await auth.check();
+//     return res.json(response).status(response.statusCode);
+//   }
+// }
 const process = {
   login: async (req, res) => {
     const auth = new Auth(req);
@@ -22,5 +22,4 @@ const process = {
 };
 export default {
   process,
-  output,
 };
