+++ app/src/routes/auth/authRoutes.js
@@ -2,33 +2,13 @@
 
 import express from "express";
 import ctrl from "./authCtrl.js"
-import jwt from "jsonwebtoken";
+import middleware from "./authMiddleware.js";
+
 
 const router = express.Router();
 
-router.post("/login", (req, res, next) => {
-  let token = req.headers['authorization'];
-  if (!token) {
-    const response = { error: 'Bad Request', message: "토큰이 없습니다.", statusCode: 400 }
-    return res.status(response.statusCode).json(response);
-  }
-  const [type, clientAccessToken] = token.split(' ');
-  try {
-    let payload = jwt.verify(clientAccessToken, process.env.ACCESS_TOKEN_SECRET);
-    console.log('토큰 인증 성공', payload)
-    req.user = { ...payload };
-    console.log(req.user)
-    // res.json({ message: "인증 성공" });
-    next();
-  } catch (err) {
-    console.log("유효하지 않은 토큰입니다.");
-    res.status(401).json({ error: 'Unauthorized', message: err.message, });
-    next(err)
-  }
-},
-  ctrl.process.login
-);
+router.post("/login", middleware.tokenProcess.accessToken, ctrl.process.login);//여기서 미들웨어는 토큰 검증이 되는지 확인하기위해 사용함(로그인할때 액세스토큰 필요x)
 router.post("/access", ctrl.process.access);
 
-router.get("/check", ctrl.output.check);
+// router.get("/check", ctrl.output.check);
 export default router;
